Reimplement host/service deletion as commands
service deletions for services on hostgroups is hard to get, since we
in naemon doesn't know if it is a hostgroup or host service.

So we need to test that against nacoma before removal, and quite early
on ask if it is ok in the future to remove them.

The simplest way of solving the problem.

Change-Id: I349ab43be5c585d071c1d768e2609c15c26f09f6
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>