Let objects have a readable nice name
When selecting objects from dropdowns, and when submitting commands and
such, a nicer name than host;description or 313;1 for downtimes is
quite nice.

Add get_readable_name to all objects generically, and make that an
alias for get_key, so all objects have a readable name. Override that
function for objects that want something special.

downtimes, comments and services is a great start

Change-Id: I9dd8ff4df51c1de13334d30feba8975af3052272
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>