ListView Multiaction: Use .attr('data-stuff') instead of .data('stuff')
.data() can handle more data types and objects, but is stored within
jQuery. Appearently jquery loses track of that data when changing
filter within a table, so multi actions stops working.

Usin .attr('data-stuff') can only handle strings, but is stored as an
actual attribute in the dom. Thus jquery doesn't need to keep track of
it, so it works much better.

Change-Id: I1426c75159f86ef4d132ed1670e2e655540f17be
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>