ListView: Make mutli-select checkboxes available for all tables
at least for those who have commands.

The multi action commands is quite generic nowadays, which means as
long as an object presents commands, those are selectable in a menu.
But to use a command, the checkboxes needs to be available too.

So try to present the checkboxes for all tables, and if there are no
commands for the table, let the column decide that it shouldn't be
visible.

Change-Id: I58d11936223546c78b96cf735505b9a564b981a9
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>