Resolve races in multi-action tests
Capybara failed to click on the correct link sometimes in the
mutli-aciton feature steps. The menu zooms out in 100ms, but capybara
is sometimes faster than that, and places its click incorrectly, thus
missing the correct link.

This should sleep for 10 times as long as the menu takes to zoom out,
which should be enough to make it fully animated.

Should make the builds more stable

Change-Id: I1fe647cf6360ccc266d009075c3dc3df30d52800
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>