xss: Escape menu items
Since the new menu is dynamic and based on user input, we cannot trust
it.

However, it seems that the branding module must be allowed to XSS us,
since the brand::get() interface returns an HTML string instead of
getting the Menu passed to it and setting its values on it, thus making
itself a special case.

This fixes MON-8361

Change-Id: I318e9054aad2225150c60476b8b2111476a49dd9
Signed-off-by: Carl Helmertz <chelmertz@op5.com>