Refactor widget handling
Widgets were added using database access in the post step in the rpm
file. It's much better and cleaner to use dropdirs, and it shouldn't be
that slow.

But to do that, almost every widget handling javascript needs to be
rewritten. So do it the easy way, rewrite the Ninja_Widget_Model to use
ORM. Also add support for metadata within the widget itself to contain
better information about nicer names and if they can be instanced
directly.

Resolves MON-5904

Change-Id: Ia6e1bce58c15d7b663b5dec69e75bb00ce7ce338
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>