ORM: Convert constructor from storage backend to factory
We want to be able to create objects without filling in data from
storage backend. The storage backend is just a special case if we
actually fetch data. Thus, convert it to a factory method instead of
keeping that logic in the constructor

It wasn't a problem earlier, when ORM was read only. But we need write
support in the future (we already have it though), so we need to keep
the constructor clean.

Resolves MON-8419

Change-Id: I67115de3d27fcb882fb9a715b4c143cd34917576
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>