Old_Timeperiod_Model: keep it a while longer, but use livestatus
We have implemented timeperiod schedule information in livestatus,
which is much better to use to get timperiod information out, since
merlin needs to be simplified, and we need to remove ocimp, which was
used to populate the database for timeperiod information.

Old_Timperiod_Model is in fact old, and should be converted to ORM, but
for now, use least-impact-change just to remove the need of merlin
status database, so we can strip ocimp out of merlin.

Also, there were a case of undefined behaviour, having duplicate
entries in a timeperiod. That's something for naemon to resolve, so we
shouldn't keep that here. The behaviour is not defined in documentation

Resolves old_timeperiod_model part of ninja part of MON-8443

Change-Id: Ib7f826405dc3144859ea14ac7c7926f19abdbf30
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>