Tests: Convert report test case definitions to PHP arrays
They were defined as a DSL language, which was just 1:1 mapped to a php
array (which broke out a couple of variables during parsing though).
But it isn't possible to have arrays, just objects, in the langauge,
which makes it hard to actually create timeperiod models as they are
returned from livestatus.

Therefore, just remove the DSL parser, and convert those test cases to
arrays. No information have been changed, since conversion has been
done with the original parser and exported to php arrays

Change-Id: I4f8ccb9ba519af8647277ea0aed046e16cdd7768
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>