summary: Allow mapping of undetermined filter_status
Remapping the undetermined state of SLA reports would not get saved.
Viewing a report in the GUI worked alright once but values for remapped
statuses were not saved.

op5reports has two subreports, one SLA and one Summary, and the Summary
one did not allow the undetermined state to be remapped.

Summary reports was the only report type that unset the
undetermined/unknown/pending state. I am unsure of why this was a good
idea in the first place. The same Summary_options class even has logic
(see update_value()) that uses the magic -1 option ('undetermined' in
the GUI, 'pending' in the code).

This fixes MON-8220

Change-Id: I2d31284377feddef2a1609deda527658dc4a9531
Signed-off-by: Carl Helmertz <chelmertz@op5.com>