schedule_reports: Don't expect $template->menu to be set from CLI env
When scheduled reports were to be sent out, we got PHP errors in the
op5-ninja log because of syntax errors. The menu template affected in
this patch assumed that $menu was always set, but it obviously (in
hindsight) is not always set, not if we're coming from generate_pdf() in
the reports module for example, or any other CLI request that renders a
view that contains the 'main' template of ninja.

This fixes MON-8465 and MON-8466

Change-Id: I39951ec36decf0c2b9acaa36fcc348cba369925f
Signed-off-by: Carl Helmertz <chelmertz@op5.com>