Introduce exec() replacement: the proc helper
We want to be able to separate stderr and stdout, exec() does not allow
us to do that. proc_open() is a builtin php function but it requires a
lot of boilerplate to setup.

proc::open() creates a simple syntax[1] on top of proc_open() which
fulfills our need of separating the output streams' contents.

    1: modelled after Python's subprocess.call()

Change-Id: I39f6252aaf01d0580e16fd990c1bdfe26cdb1b81
Signed-off-by: Carl Helmertz <chelmertz@op5.com>