spec: chown /etc/op5/*yml to monitor user, fixes backup
It's impossible to create the kind of Backup you reach through "Manage
-> Backup/restore", if any files are unreadable by the PHP user.
Permission checking is very hard to control in runtime. We could stat
every file and exclude the non-readable ones but that could result in
data-loss (a backup is supposed to be 'complete'). Instead, we keep the
error message that says 'something went wrong' and deal with a common
case when upgrading from older Monitor versions - configuration files
created by ourselves in order to keep config-noreplace together with not
going down the .rpmsave or .rpmnew path.

Fixes backend part of MON-7208

Change-Id: If651bc99233d3c0a999e4fc69ebae5edf4791adf
Signed-off-by: Carl Helmertz <chelmertz@op5.com>