backup: Replace command execution with proc::open()
We know that tar output errors on stderr and that Naemon outputs
(config verification) errors on stdout, which makes us able to decide
when to show what, to the user. This matters greatly when a backup
fails, since tar[1] outputs all successful operations on stdout, which
PHP's exec() mixes with the much more important stderr. Stderr of the
failed backup job is something the user can react to (chmod/chown),
instead of being fed with the technically correct but ultimately useless
information in the GUI.

    1: we cheat a little bit by knowing that op5/backup/backup proxies
       tar

Fixes UX part of MON-7208

Change-Id: I50ae67e07e50c5d8932e91154e7290695fdc0761
Signed-off-by: Carl Helmertz <chelmertz@op5.com>