monitoring-menu-hook: Don't expect order of hooks
$menu->get('not yet existing') will return NULL. So, depending on the
order of callbacks for the event ninja.menu.setup, menu items might or
might not have been loaded. Since we can add 'not yet existing.hello'
before 'not yet existing', we are fine, so let's do that. 'not yet
existing' will overwrite that node later on, but keep the child nodes.

Moral of the story: do not use Menu_Model::get()->set(), you have to
check the result of get() before using set(), because it might be NULL.

Change-Id: I2cf3bdf83ab372575f81f35abb5fefd8a5841422
Signed-off-by: Carl Helmertz <chelmertz@op5.com>