search: Remove caching option
The caching fails at colon-filter syntax due
to it asserting that anything entered after
h:mon will not produce any more hits, which is sensible.
However, the syntax allows for AND/OR which means
it could be followed by AND s:Cron which should
produce service results, but didn't.

This fixes MON-5113

Change-Id: If0f2ba9aec337ff864a6c4f41577869e41acdc0f
Signed-off-by: Tobias Sjondin <tsjondin@op5.com>