Implement datasource-drivers for better decoupling
This adds another level of abstraction between object pools and their
data source, called a driver, and an interface that defines them. This
allows us to decouple object pools from the data sources in which they
are stored, allowing us (for example) to store and retrieve monitored
objects in a RDBMS instead of LiveStatus or store log data in an
ElasticSearch instance instead of an RDBMS.

Most importantly right now, however, is that we may define our own
"PHP-native" data source, which allows us to mock object data whichever way
we want to. That's not included in this particular commit, though.

This fixes MON-8699 and MON-8697.

Change-Id: I83faffc3351beba150465effeff60ea4e00d6148
Signed-off-by: Anton Lofgren <alofgren@op5.com>