Implement a (PHP-)Native data source driver
This comes with a filter builder visitor which produces native
operations on the statements it visits.

Together, these components allow us to use native PHP data structures to
expose ORM data. This is particularly useful when mocking, since we can
replace a table's default driver with a ORMDriverNative and inject
whatever data we like.

This fixes MON-8698.

Change-Id: I588036b6cd451306eba47718b95c5ac3907a6947
Signed-off-by: Anton Lofgren <alofgren@op5.com>