test: Add data mocking interface
This patch introduces a rudimentary interface for mocking of
data drivers. It listens for a HTTP header 'X-op5-mock' which should
have a (server-)path as its value to a file containing the data mocking.

The file should contain a JSON array of the form:

{
  <'DriverClass1> <instance-name>': {
      '<table-name1>': {
	'field1': 'value1',
	'field2': 'value2'
      }
    }
  <'DriverClass2> <instance-name>'
  ...
}

and so on.

To make it more practical to use, the event hook mocks out op5auth and
op5MayI with very permissive variants.

This fixes MON-8701.

Change-Id: Icf31da6a2b2cf0d745e5b8c73f7dea401c8d64ae
Signed-off-by: Anton Lofgren <alofgren@op5.com>