Remove features testing Nacoma
Ninja should not test its modules, it should provide a framework of
testing those interfaces. Otherwise, a Nacoma build depends on a Ninja
build for the proper interfaces, and a Ninja build depends on Nacoma in
order to perform these tests. Remove that last dependency.

Marking features with @skip tells (our invokation) of cucumber that
those steps need to fail, adjust them a bit.

Change-Id: Ife5b60545229b73c987a2761239d7919fdff0a69
Signed-off-by: Carl Helmertz <chelmertz@op5.com>