Remove features testing Nacoma
Ninja should not test its modules, it should provide a framework of
testing those interfaces. Otherwise, a Nacoma build depends on a Ninja
build for the proper interfaces, and a Ninja build depends on Nacoma in
order to perform these tests. Remove that last dependency.

ORM_Test::test_host_config_url() tests the ninja-part of this.

Change-Id: Ife5b60545229b73c987a2761239d7919fdff0a69
Signed-off-by: Carl Helmertz <chelmertz@op5.com>