ORM: Generate factory_from_array method for objects
It often gets cumbersome to deal with backend names, since they are very
driver specific. Usually, it's not a huge problem, but for drivers like
ORMDriverNative it's very hard to write a generic solution since it
could be used to mock any old table, regardless of its original backend
driver.

Change-Id: I5b9ea877e7735e67cda96b9dfd4b09ee2c412a8c
Signed-off-by: Anton Lofgren <alofgren@op5.com>