Do not show unreliable next_check values for remotely checked objects
We cannot trust the next_check value on the master when the host/service is
checked by a poller or peer, instead of lying on the extinfo page we now print
"Remotely checked by <remote_name>". This does not change the behaviour of the
next_check value in the ORM, i.e. you may still fetch the invalid timestamp via
custom columns and the HTTP API.

This fixes MON-8812

Change-Id: I033178c1f23441c6cb3f302c1d6fb48145afa0cc
Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>