In scheduling queue, hide remotely checked items
Since we cannot determine when or if they are actually going to get checked.
Had to refactor the scheduling queue to get this to work and it is still not a
nice solution using two pools that are joined into an array. However, the view
is now neater, using all helpers at our arsenal :) As a bonus I got the
pagination to work again.

This patch also makes use of the driver mocking functionality in Ninja
for its tests. The tests could be prettier, but that's more of a
question of ORM providing sensible default values based on the type of
an object's fields.

This fixes MON-8833

Change-Id: I6b97ec2b6ffdeaa7795dc53f826e2837579c6bdc
Signed-off-by: Anton Lofgren <alofgren@op5.com>