report_options: Don't throw exception if report contains 0 objects
I tried to look at a saved report by clicking the report's name in the
SLA menu. I was met with the exception:

modules/reports/libraries/Sla_options.php [154]:
reset() expects parameter 1 to be array, null given

A "select * from saved_reports_objects where report_id = Y" (where Y is
found in the URL), yielded 0 rows. I guess that the cause is the ninja
hook that removes reports' objects when the objects are removed, an
operation that ignores if the report is left with zero objects.

Change-Id: Iabe2991872d10e0decfe64328eeecad742de291f
Signed-off-by: Carl Helmertz <chelmertz@op5.com>