Removed box shadow from widget and header
We have removed gradients and blurry grey colors from all other places, its
time for us to let go of these as well.

Also edited the margin of widgets so that the horizontal margine is the same as
the vertical, previously the horizontal margin was twice the vertical when
rendered.

This does not resolve any issue but should have been done in the last sweep of
gradient removals.

Change-Id: Ib99778dd2faf1743eeaaead1d4381b3140e93e19
Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>