janitor: Fix widget widths when divided by an uneven number
One-third of 100 is not 33 and its ugly to pretend so, make it slightly more
accurate by allowing two decimals in the width style attribute so the widgets
fall in line with the other rows widgets.

This is only a slight graphical issue that has not been reported, reporting it
is more overhead than it is worth.

Change-Id: I36a81a7a910ce6189d104b400ed4d38cddc6c68e
Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>