Drop op5cucumber dependency
This patch drops Ninja's dependency on op5cucumber. The op5cucumber
package has turned out to be a large roadblock in terms of being able to
rapidly develop tests. Since we remove the dependency, we need to
duplicate some of the steps and helpers that exist in op5cucumber.
Hopefully, we'll be able to migrate the rest of them to the module
(logger, trapper, ...) where they're actually, eventually.

Change-Id: I7b3a30ffe4dddcb9dcceb08960dcba53c7e17ce6
Signed-off-by: Anton Lofgren <alofgren@op5.com>