Add support for .env files and configurable modules
.env files are useful for injecting local settings into Ninja without
having to reconfigure apache. Emphasis on "local" here, this is not
meant for production use, but rather for use in CI environments or for
developer specific settings.

Configurable modules are useful for ensuring a consistent setup, for
example when running tests.

Both of these changes together are an important step towards being able
to reliably run cucumber tests on a developer machine/vm/container.

Change-Id: Icf6ea10d9754ce497d4f95f64424501989263237
Signed-off-by: Anton Lofgren <alofgren@op5.com>