Move the global notification string handling to object instances
Global notifications are from here on out called Notices as to
avoid confusion both in communication and development. A notice
has a type and a message, where type may be error, warning, info,
or success.

The notices created should be contained in the NoticeManager until
such a time as they should be rendered.

This fixes MON-8966

Change-Id: Ia0487153502f6a4b535c0185465bff827fcc96a7
Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>
Signed-off-by: Carl Helmertz <chelmertz@op5.com>