mock: 'After' in cuke does not have access to page.driver.browser
Store mock's state in the mock instance. We do that instead of keeping
track of the reason why the mock was creating the first place, by
checking the HTTP headers again, since there are no reference to the
headers left in the After (teardown) state of the test. Blindly relying
on the headers' existence, Ninja failed all cuke tests with an error:

undefined method `headers' for
 #<Capybara::Poltergeist::Driver:0x00000002ba5728> (NoMethodError)
/mnt/git/features/step_definitions/mocked_configuration.rb:26:in `After'

Change-Id: I75608a33aa29899ac732e15371b7c451d85adb6f
Signed-off-by: Carl Helmertz <chelmertz@op5.com>