process_check_result: Use sane defaults for perf_data
Let `null`, not `false`, signify an unset value. This is not a boolean
context, so let's not use booleans to communicate that we have no
perf_data.

This fixes an issue where check results with no perf data instead ended
up with empty perf data, meaning that the command was appended with a
pipe delimiter. This is a problem in particular when the perf_data
already exists in the $plugin_output, since it would then be overridden
by the bogus empty perf_data.

This fixes MON-8848.

Change-Id: Ic4eeb8fc06b7c282c0b3cc3d26595c6aae4a203d
Signed-off-by: Anton Lofgren <alofgren@op5.com>