reports: Place icons in toolbar
Cucumber had a problem with finding the icons on the report pages, for
example it couldn't click on "Show as pdf" since that element was
covered with something. The Cucumber screenshot was useless since it
does not include images.

This refactoring moves logic out of the view and into the controller,
although it could be split up more in the future.

This makes way for testing being able to test MON-8113 properly.

Change-Id: I1777e1e3067bf04b58a72a42adf69a26c70961da
Signed-off-by: Carl Helmertz <chelmertz@op5.com>