Make widget loading more robust
If a widget fails to load/instantiate or leaks an exception upon
indexing, it would previously take down Ninja in its entirety.

This patch makes the widget loading a bit more robust by checking for
errors and replacing dead widgets with a placeholder Dead_Widget which
communicates the error to the user.

Of course, widgets are still welcome to handle their own exceptions,
this just serves as an additional safeguard for widgets that don't.

To enable running the widget tests locally, we also had to change a few
cucumber helper functions, which is why there are a number of seemingly
unrelated test changes, as well.

This fixes MON-8307.

Change-Id: I5651c1e0dd2cbb3eebe4952515b94c6eb7dcee9c
Signed-off-by: Anton Lofgren <alofgren@op5.com>