notices: Disallow duplicate messages (same content)
The user ends up with duplicate messages, which is bad. We don't check
on before hand if we already added a similar message. Checking for that
would lead to a lot of extra code all over the place. Instead, we treat
the notice manager as a "set" of notices (that is, no duplicates).

Case in point: we use the method _verify_access() sequentially to test
one mayi action at a time. If each call adds a license warning, we
make checking for notices O(1) over O(N).

Highest prio wins.

Change-Id: Id705fac279cc05393051ee0523e73a46830e6bb9
Signed-off-by: Carl Helmertz <chelmertz@op5.com>