Make is_subset available to rest of system & fix a few bugs
This patch moves the is_subset function from the mayi auth hook to a
static member of the MayI class, in order to allow other parts of the
system to call it. I also took the liberty to reduce code duplication,
document it & write a few test cases for this rather important function
and fix a few minor bugs.

Change-Id: If2f5129ae3466d2092d4a05b438c717e30ac0b03
Signed-off-by: Anton Lofgren <alofgren@op5.com>