Add support for more granular mocking of op5MayI and op5Auth
In order to be able to write tests such as "Given I'm not authorized to
do A, I should see error message M", we need to be able to easily mock
authorization points and namespace set permissions. This patch
implements that, by expanding the existing MockAuth and MockMayI
classes.

This change requires (well, not really, but it makes the implementation
a lot simpler) us to change the X-op5-mock JSON structure a bit as well.

Change-Id: I7d2eaeccb0e01d653a65cc6284b2e9f5aab066b8
Signed-off-by: Anton Lofgren <alofgren@op5.com>