Don't be cryptic with redirection intent
The mayi hook ACL keeps a 'ninja.auth:login' namespace which is
accepted when the user is not authorized and rejected otherwise.

Among other things, it is used here to determine, for a user that has
been denied access to "something", whether we should redirect her to the
login page (when she is not logged in), or simply show the regular
"You're not allowed to be here" blurb.

This works fine in practice, but makes the intent in this instance a bit
too cryptic and unexpected. It also makes it hard to reason about which
namespaces I need to deny in my test cases when testing permissions &
restrictions, so I'm instead opting to be very explicit about our intent
here, and leaving the namespace out of the equation.

Change-Id: Ie822f6e3818340f1eccd694446c18cdf1c584761
Signed-off-by: Anton Lofgren <alofgren@op5.com>