tac: Don't show error notice when AJAX aborted through refresh
When an oustanding request, such as a widget refreshing, was aborted,
the TAC would show an error notice like:

"Error calling on_refresh"

Since the user themselves aborted the refresh (probably by refreshing
the page in their browser), it is not necessary to inform them of this.

It is also not an error.

This fixes MON-9115.

Change-Id: I6100181379716b850949f37ae428e84fba53b223
Signed-off-by: Anton Lofgren <alofgren@op5.com>