widget/css: Avoid overflow for widget content
Widgets that don't take care of content overflow by themselves, such as
the DeadWidget widget are liable to overflow the content div, causing
the appearance of an ugly horizontal scroll-bar.

This patch fixes that by setting the word-wrap rule.

Note that we use word-wrap rather than overflow-wrap since the former
was renamed to the latter in CSS3, and we have no reason to provoke
incompatibilities here.

Unfortunately, despite my efforts I have been unable to find a way to
test this behaviour with Capybara & Poltergeist, so manual/ocular
verification needs to be done.

This fixes MON-9133.

Signed-off-by: Anton Lofgren <alofgren@op5.com>