auth: op5auth::get_user() had errors from previous refactoring
There was a reference to an undefined variable, $auth_method, in
get_user().

I also added tests to not let this be repeated. The tests were placed in
the Apache auth driver test suite because I reached the bad code path
(and errored out with "undefined variable") during manual testing of the
Apache auth driver, more specifically, by running HTTP API tests.

Since op5auth::__construct() depends (instead of taking it as a
constructor argument, sigh) on op5config, we must access the
might-be-mocked version of op5config, so let's use the op5objstore.

This is needed to fix the HTTP-API issue MON-9143

Signed-off-by: Carl Helmertz <chelmertz@op5.com>