Work around Kohanas poor events in ninja.setup
Get a copy of the event data to pass to the sub-events
of ninja.setup so that the Event data is not lost after
ninja.setup.user

This fixes current ninja master build issue of
"Uninitialized string offset: 0" which happend due to:

The Event ninja.setup recieves Event::$data (note the
static public property).

This event in turn fires two "sub-events" as these setup
steps are required to run in sequence, due to the
contact creation requiring the user.

However, the Kohana Event system does no like this one
bit, the Event::$data is "cleared" (i.e. set to a
reference to a string O.o) after the firing of an event,
in this case ninja.setup.user, before running
ninja.setup.contact- So when we are in
ninja.setup.contact the Event::$data is an empty string
reference, that we attempt to use as an array with a
string index, which is cast to int 0, badda bing badda
bam we have "Uninitialized string offset: 0".

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>