mayi (_verify_access()) should be able to deny login attempt
We correctly caught the "mayi says you're not allowed, and you're not
logged in, so, I'll show you the door"-path. The thing is, we need to
exit() after calling header("Location: X"), otherwise the execution will
continue.

This was noticed by manual testing on head664: someone tried to login,
the license check was blocking the login, but the redirect never kicked
in. That chain of events led to a semi-logged in state -- everything in
the GUI was a "bit off" and there were a few more red boxes with cryptic
error messages than there should be.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>