A base class cannot rely on methods to exist in inherited classes
Simplified view of our controller inheritance:

                  Base
                 /    \
             Ninja    Chromeless

_verify_access() is defined in Base, and calls add_print_notification()
which is defined in Ninja. This is not ideal for those controllers
inheriting Chromeless and still relying on being able to call
_verify_access().

This manifested itself in Kohana catching the call to the not-found
add_print_notification(), thinking that it is an url (because it's a
method on a controller), and rendering a 404. *But* the GUI reports
that, for example, auth/login was not found even if that url and method
exists. That is just because auth/login (in this case) was the method
calling _verify_access(). End result: very confusing output.

Composition over inheritance, say it with me!

This all manifested itself by not being able to login to Ninja. I added
debug calls to every known location that triggers 404s, and decided to
keep them in there since they might help someone else. They shouldn't
spam logs in any well functioning system since 404s should be a) rare;
b) reported, dealt with or simply ignored.

This fixes part of MON-9163 and allows us to login to a Ninja system
again.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>