Remove PNP features, they have been moved into PNP
Our cuke steps that test a Ninja module's logic should reside in that
Ninja module, not in the parent project that is Ninja.

The features were explicitly marked @unreliable, which means that they
do not fail builds. Since their import into PNP, they are tested with
each build.

This was part of the refactoring for MON-9182