Fix the get_metadata method of auth
For specific field requests, if no auth_modules exist, the variable $var
will not be set, and even if $var is set it will likely not be the result
you wanted, what you want with specific calls is the member $field from the
metadata array.

So the interface has been fixed, it's still an odd interface and will likely not do what is generally expected, but I've added some unit tests to further ensure the behaviour.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>