Move ninja settings into ORM
So that they may be mocked in tests, in addition it is
quite likely this has removed some unknown SQL injections
provided by the Ninja_Settings_Model class.

This changes no existing public interface.

The reason for doing this is so that I can mock settings
when testing the Wizard.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>