Create a separate Kohana compliant autoloader
Which may be used from other repos that depend on ninja classes. Note that this
will not execute hooks within ninja only give access to ninja classes.

The autoloader will not pollute the namespace of any other script it is
loaded into as it executes in a closure once required.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>