Re-add Kohana's internal autoloader
Kohana's autoloader was ripped out into a separate autoloader some
commits back. This commit reverts the ripping out part, but keeps the
ripped out part.  That's because PNP is installed as a separate Kohana
installation, just without the central "system" folder. That folder
refers to/hijacks/reuses Ninja's system folder instead, just like that
-- meaning: changes in Ninja's system reflects on other projects.

If we keep the system folder as it was before, PNP can continue using
it. The long term plan must be to treat the graphing part of Monitor
with a more well-defined and less intrusive interface.

This fixes parts of MON-9182

Signed-off-by: Carl Helmertz <chelmertz@op5.com>