Display nice error messages when Livestatus is down
ORMDriverExceptions are now handled in a general way. General
for everything except listviews, which needs their own
handling since they are loaded from frontend.

When Livestatus is down it retries to connect to it three times
before giving up.

Added classes to trigger an ORMDriverException to be used in
Cucumber tests.

This fixes MON-9075

Signed-off-by: Joachim Ekstrand <jekstrand@op5.com>