test: Remove deprecated Naemon configuration variables
Each time Naemon is restarted, the log is filled with messages such as
this:

Warning: daemon_dumps_core is deprecated and will be removed. Use sy...
Warning: auto_rescheduling_window is deprecated and will be removed....
Warning: auto_rescheduling_interval is deprecated and will be remove...
Warning: auto_reschedule_checks is deprecated and will be removed. A...
Warning: max_host_check_spread is deprecated and will be removed. Ho...
Warning: host_inter_check_delay_method is deprecated and will be rem...
Warning: service_interleave_factor is deprecated and will be removed...
Warning: max_service_check_spread is deprecated and will be removed....
Warning: service_inter_check_delay_method is deprecated and will be ...
Warning: log_rotation_method is deprecated and will be removed. Logs...
Warning: nagios_group is deprecated and will be removed. Naemon is c...
Warning: nagios_user is deprecated and will be removed. Naemon is co...
Warning: check_result_path is deprecated and will be                ...

It hides the things you are actually looking for while examining the
logs, especially when we build Ninja in our CI environment.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>