Model tac in Dashboard_Model
Since tac needs to handle logic related to the tac itself, not only to
its widgets, we should model it as such. Doing so means that we can
decouple the tac from its user, and easily implement multiple tacs and
sharable tacs in the future.

It also means that we in the future have a place to put methods for
serialize/deserialize entire tacs, or reorder widgets, without being
coupled to a certain user. It makes it much easier to test.

This is a commit squashed from multiple WIP-commits from several people

Resolves most of MON-9265

Signed-off-by: Joachim Ekstrand <jekstrand@op5.com>
Signed-off-by: Robin Hagman <rhagman@op5.com>
Signed-off-by: Carl Helmertz <chelmertz@op5.com>
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>