Use Kohana include paths to resolve manifests
This has been one of those details which just made loading of manifest
files just semi-working in unit tests of external libraries, since I was
lazy when implementing this function long ago.

Since Kohana have the concept of include paths, which is a list of
"modules" (not only modules, but also the application and system
directory), tests in external programs mangles that include path to
inject the code from git to the auto loader. But manifest files ignored
that and listed modules by itself.

This patch uses the include path instead, and also stores the manifest
cache in a singleton class managed by op5objstore, so it will be cleared
between tests, so mangling of that cache will be possible between tests,
to make test cases more isolated.

Identified and part of MON-8735 (but actually quite generic)

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>