Permission quarks: First implementaiton
This is concept of permission handling, where each user generates a
regexp to match against a comma seperated list of quark id:s for each
object that should have permission.

The list should be prefixed and suffixed with an additional comma, to
simplify the regexps, for exmaple:

,2,5,6,

should give access to quark id 2, 5 and 6.

A user which have quarks 2,3,4 generates the regexp: ,(2|3|4), which
matches.

quarks is a mapping table between permission objects (for exmaple "user
max", "group admins") and a unique id

Part of MON-9337

Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>