Dashboard_Model: Add {export,import}_array() functions
They do thoroughly unsurprising things, given their names.
->import_array() replaces the current tactical overview
with new settings in an array of the format
  array('dashboard' => ..., 'widgets' => array(...));

This relates to MON-9346, but doesn't completely fix it
due to this little appendix in the description part of
the issue: "and also update tac to create the first
dashboard from a ninja config file.", which I feel
should be a separate issue anyways.

Signed-off-by: Andreas Ericsson <ae@op5.se>