widget: replace get_option_groups() with Fieldset_Model
The view became so damn messy when it had to keep track of all options that
were loosely related. The options also had to be in the same order, making
option_groups a very fragile solution.

I'm not rebasing this since it is an API break and I want it to be easily
revertable.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>