Widgets: big number and state summary height
Update the min-height of state summary widget so that it matches that of
big number widget. While at it, change the space indented style to use
tab indent.

Remove the "-" in big number widget when no objects are found and move
the display_type to the center of the widget.

This fixes MON-9399 and makes MON-9381 look a bit better.

Signed-off-by: Robin Hagman <robin.hagman@op5.com>