mayi: Let {host,service}_edit_all delete downtimes/comments etc again
The fix for MON-9376 in 868edb4ac2d was a bit overzealous. The fix
for that issue correctly removed the monitor.monitoring.hosts:delete
right for the host_edit_all permission, but it also removed related
rights such as monitor.monitoring.comments:delete. That meant that a
user could for example not delete any comments or downtimes
anylonger, although they previously could do that with only
host_edit_all. There is no reason to change that behaviour.

This fixes MON-9413.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>