config: Don't require specifying a page when getting a setting
This function was only ever used with a '*' wildcard for $page, to
indicate "any page". This patch deprecates that parameter, along with
the $save parameter, since that has no place in a get() function.

As a side effect, this fixes MON-8565, since the code that was supposed
to retrieve the default command settings didn't specify the wildcard,
but null.

Signed-off-by: Anton Lofgren <alofgren@op5.com>