ci: Only put logs in /mnt/logs if we are in a ci environment
If I want to do some testing locally, I want to install monitor-ninja-test,
which gives me both the correct dependencies, and all of the actual test files.
I *do not* want to modify the running system's main configuration, and I
especially do not want to redirect the log output into a folder that is
non-existent on my machine. That redirection is only ever useful in a CI
environment, which is controlled through ci_config.yml, so let's put the
configuration there.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>