hooks: Only execute hooks that are PHP files
A bit of a security issue: if we `include` files in a PHP context, they must
contain PHP code. The regular way to do that is to name files x.php.

I discovered this by applying a patch with a wonky history, ending up with
files named *.php.rej in a specific module's hooks folder, which made Kohana
load it and crash afterwards, since it spit out text where there should be PHP
logic.

Signed-off-by: Carl Helmertz <chelmertz@op5.com>