queryhandler: Don't use backtracking regex to parse kvvec
For long strings, a backtracking regular expression is unsuitable for
parsing kvvecs, since the backtracking will eventually trigger a
PREG_BACKTRACK_LIMIT_ERROR.

Instead, we implement a simple parser that doesn't have this problem.

This patch also adds a bunch of tests for the kvvec2array function,
including one that triggers bugs MON-7359 and MON-9046.

This fixes MON-7359 and MON-9046.

Signed-off-by: Anton Lofgren <alofgren@op5.com>