mayi: Don't overwrite overlapping environment subtrees
This patch fixes a bug in the MayI environment handling which would
cause an actor's subtree whose supertree was already included in the
environment to be overwritten by subsequently added actor's subtrees
belonging to the same supertree, when the latter was a "raw" value
(such as an integer).

As a concrete example, this fixes a bug where we would erroneously drop
constraints for the monitor.monitoring subtree when the
monitor.installation_time environment was included. This in turn would
lead to misleading results when querying constraints on the
monitor.monitoring subtree, such as "May I add more hosts?".

This fixes MON-8975.

Signed-off-by: Anton Lofgren <alofgren@op5.com>