Refine widget dragging, move to flex layout
Move the dashboard to a flex-box layout, flex-box is currently only supported
under prefixes but is available for IE10, chrome and firefox. The effect is
that it makes it far easier to set the width of the widget columns and moving
widgets between columns (within the same row especially).

Add the proper options to jquery-ui sortable (remove containment) to make
easywidgets a little less bad, you can now drag from the middle of a 3-col
widget, into a 3-column row to the far left, something previously not working.

Signed-off-by: Tobias Sjöndin <tsjondin@op5.com>